# [perf] Profile possibly slow functions