## Summary

Profile a few functions. This shows any function that takes more than 1 ms. 

Q: does trace have this functionality? It looks like it does time events but we have to explicitly enable it using the `trace` flag? 

## How was it tested?



```bash
export DEVBOX_PRINT_EXEC_TIME=1

devbox add curl
...
Exec times over 1ms:
"syncPackagesToProfile" took 343.118292ms
"nixEnv" took 128.018333ms
"nixEnv" took 128.219083ms
"CreateWrappers" took 257.527167ms
"ensurePackagesAreInstalled" took 768.378084ms
"CreateWrappers" took 2.56ms
"devbox add curl" took 1.21434s
```